Unify test_package_install & test_package_json_rewrite
